+++ src/crawlers/basic_crawler.js
@@ -3,11 +3,12 @@ import _ from 'underscore';
 import log from 'apify-shared/log';
 import { ACTOR_EVENT_NAMES } from 'apify-shared/consts';
 import { checkParamPrototypeOrThrow } from 'apify-shared/utilities';
-import AutoscaledPool from './autoscaling/autoscaled_pool';
-import { RequestList } from './request_list';
-import { RequestQueue, RequestQueueLocal } from './request_queue';
-import events from './events';
-import { addTimeoutToPromise } from './utils';
+import AutoscaledPool from '../autoscaling/autoscaled_pool';
+import { RequestList } from '../request_list';
+import { RequestQueue, RequestQueueLocal } from '../request_queue';
+import events from '../events';
+import { addTimeoutToPromise } from '../utils';
+import Statistics from './statistics';
 
 /**
  * Since there's no set number of seconds before the container is terminated after
@@ -192,6 +193,7 @@ class BasicCrawler {
         this.handleFailedRequestFunction = handleFailedRequestFunction;
         this.maxRequestRetries = maxRequestRetries;
         this.handledRequestsCount = 0;
+        this.stats = new Statistics({ logMessage: 'Crawler request statistics:' });
 
         let shouldLogMaxPagesExceeded = true;
         const isMaxPagesExceeded = () => maxRequestsPerCrawl && maxRequestsPerCrawl <= this.handledRequestsCount;
@@ -256,7 +258,14 @@ class BasicCrawler {
         await this._loadHandledRequestCount();
         this.autoscaledPool = new AutoscaledPool(this.autoscaledPoolOptions);
         this.isRunningPromise = this.autoscaledPool.run();
-        await this.isRunningPromise;
+        this.stats.startLogging();
+        try {
+            await this.isRunningPromise;
+        } finally {
+            this.stats.stopLogging();
+            const finalStats = this.stats.getCurrent();
+            log.info('Crawler final request statistics:', finalStats);
+        }
     }
 
     async _pauseOnMigration() {
@@ -328,6 +337,8 @@ class BasicCrawler {
         // Reset loadedUrl so an old one is not carried over to retries.
         request.loadedUrl = null;
 
+        const statisticsId = request.id || request.url;
+        this.stats.startJob(statisticsId);
         try {
             await addTimeoutToPromise(
                 this.handleRequestFunction({ request, autoscaledPool: this.autoscaledPool }),
@@ -335,6 +346,7 @@ class BasicCrawler {
                 'BasicCrawler: handleRequestFunction timed out.',
             );
             await source.markRequestHandled(request);
+            this.stats.finishJob(statisticsId);
             this.handledRequestsCount++;
         } catch (err) {
             try {
@@ -408,6 +420,7 @@ class BasicCrawler {
         // Mark the request as failed and do not retry.
         this.handledRequestsCount++;
         await source.markRequestHandled(request);
+        this.stats.failJob(request.id || request.url);
         return this.handleFailedRequestFunction({ request, error }); // This function prints an error message.
     }