+++ src/puppeteer.js
@@ -242,6 +242,8 @@ export const launchPuppeteer = async (options = {}) => {
         optsCopy.args.push(`--user-agent=${userAgent}`);
     }
 
+    if (optsCopy.pipe !== false) optsCopy.pipe = true;
+
     let browser;
     if (optsCopy.proxyUrl) {
         browser = await launchPuppeteerWithProxy(puppeteer, optsCopy);